Bugfix/rtc naive not using instant (#151)
* wip naive use instant

* update write impl to use flowing clock

* add test using ticking clock

+++ gb-rtc/src/constant.rs
@@ -1,5 +1,5 @@
-pub const MINUTE: u32 = 60;
-pub const HOUR: u32 = 60 * MINUTE;
-pub const DAY: u32 = 24 * HOUR;
-pub const MAX_DAYS: u32 = 0x1FF;
-pub const MAX_TIME: u32 = MAX_DAYS * DAY;
+pub const MINUTE: u64 = 60;
+pub const HOUR: u64 = 60 * MINUTE;
+pub const DAY: u64 = 24 * HOUR;
+pub const MAX_DAYS: u64 = 0x1FF;
+pub const MAX_TIME: u64 = MAX_DAYS * DAY;
